+++ lib/core/facets/enumerable/hinge.rb
@@ -9,10 +9,9 @@ module Enumerable
   #
   # TODO: Enumerable#hinge will get a new name.
   #
-  def hinge(init={})
-    h = init
-    each{ |v| yield(h,v) }
-    h
+  def hinge(init={}, &block)
+    warn "Enumerable#hinge is deprecated. Use Enumerable#each_with_object instead (note: block args are reversed).", uplevel: 1
+    each_with_object(init) { |e, obj| block.call(obj, e) }
   end
 
 end