+++ lib/core/facets/enumerable/uniq_by.rb
@@ -1,13 +1,9 @@
 module Enumerable
 
-  # Like #uniq, but determines uniqueness based on a given block.
-  #
-  #   (-5..5).to_a.uniq_by {|i| i*i }
-  #   #=> [-5, -4, -3, -2, -1, 0]
-  #
-  def uniq_by #:yield:
-    h = {}; inject([]) {|a,x| h[yield(x)] ||= a << x}
+  # @deprecated Use Enumerable#uniq with a block instead (Ruby 1.9.2+).
+  def uniq_by(&block)
+    warn "Enumerable#uniq_by is deprecated. Use Enumerable#uniq(&block) instead.", uplevel: 1
+    uniq(&block)
   end
 
 end
-